feat/promote package (#40)
Adding a promote script to promote packages from beta to upstream. This
will be a manual process to prevent any accidental publishing of beta
features. Promoting is copying a package from the node-collector-beta
repository to the node-collector repository. This is preferred over
pushing directly to node-collector where discrepancies in the process
can cause variants in the package which is problematic. Promoting
ensures identical packages are pushed to upstream without discrepancies.

Usage: VERSION=v0.0.1 packaging/scripts/promote-packagecloud.sh